+++ src/guillo-gmi/components/index.js
@@ -43,6 +43,7 @@ export * from './error_zone'
 export * from './flash'
 export * from './guillotina'
 export * from './item'
+export * from './Link'
 export * from './layout'
 export * from './login'
 export * from './modal'
